feat: ignoring legacy folder

+++ artifacts/scripts/updateFEIndexHtml.sh
@@ -2,7 +2,7 @@ if [[ -z "${FE_ROUTE}" ]]; then
     echo "No custom route exist" 
     # Loop through all folders in packages directory
     for package_dir in $TOMCAT_HOME/webapps/SemossWeb/packages/*/; do
-        if [[ -d "$package_dir" && -f "$package_dir/dist/index.html" ]]; then
+        if [[ -d "$package_dir" && -f "$package_dir/dist/index.html" && "$(basename "$package_dir")" != "legacy" ]]; then
             echo "Processing $(basename "$package_dir")"
             sed -i "/<head>/a\        <script id=\"semoss-env\" type=\"application/json\">{\"MODULE\":\"Monolith\"}</script>" "$package_dir/dist/index.html"
         fi
@@ -10,7 +10,7 @@ if [[ -z "${FE_ROUTE}" ]]; then
 else 
     # Loop through all folders in packages directory
     for package_dir in $TOMCAT_HOME/webapps/SemossWeb/packages/*/; do
-        if [[ -d "$package_dir" && -f "$package_dir/dist/index.html" ]]; then
+        if [[ -d "$package_dir" && -f "$package_dir/dist/index.html" && "$(basename "$package_dir")" != "legacy" ]]; then
             echo "Processing $(basename "$package_dir")"
             sed -i "/<head>/a\        <script id=\"semoss-env\" type=\"application/json\">{\"MODULE\":\"$FE_ROUTE\"}</script>" "$package_dir/dist/index.html"
         fi
